# backend/src/main/java/com/planifi/backend/api/dto/CreateApiKeyRequest.java
package com.planifi.backend.api.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record CreateApiKeyRequest(
        @NotBlank
        @Size(max = 100)
        String name
) {
}