# docs/openapi/planifi-api-v1.yaml
openapi: 3.0.3
info:
  title: Planifi REST API
  description: Contrato REST v1 para operaciones básicas de gastos.
  version: v1
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v1
    description: Base path para la API pública v1
paths:
  /auth/register:
    post:
      summary: Registrar usuario
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserRequest"
      responses:
        "201":
          description: Usuario registrado y token emitido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"
  /auth/login:
    post:
      summary: Autenticar usuario
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Token JWT válido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api-keys:
    post:
      summary: Crear API key para MCP
      operationId: createApiKey
      security:
        - BearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: Clave idempotente única para reintentos seguros.
          schema:
            type: string
            minLength: 8
            maxLength: 255
        - name: correlation-id
          in: header
          required: false
          description: ID de correlación propagado end-to-end.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyRequest"
      responses:
        "201":
          description: API key creada (secreto visible una sola vez)
          headers:
            correlation-id:
              description: ID de correlación propagado en toda la petición
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeySecret"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api-keys/{apiKeyId}/rotate:
    post:
      summary: Rotar API key
      operationId: rotateApiKey
      security:
        - BearerAuth: []
      parameters:
        - name: apiKeyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: true
          description: Clave idempotente única para reintentos seguros.
          schema:
            type: string
            minLength: 8
            maxLength: 255
        - name: correlation-id
          in: header
          required: false
          description: ID de correlación propagado end-to-end.
          schema:
            type: string
      responses:
        "201":
          description: API key rotada (nuevo secreto visible una sola vez)
          headers:
            correlation-id:
              description: ID de correlación propagado en toda la petición
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeySecret"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api-keys/{apiKeyId}/revoke:
    post:
      summary: Revocar API key
      operationId: revokeApiKey
      security:
        - BearerAuth: []
      parameters:
        - name: apiKeyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: true
          description: Clave idempotente única para reintentos seguros.
          schema:
            type: string
            minLength: 8
            maxLength: 255
        - name: correlation-id
          in: header
          required: false
          description: ID de correlación propagado end-to-end.
          schema:
            type: string
      responses:
        "204":
          description: API key revocada
          headers:
            correlation-id:
              description: ID de correlación propagado en toda la petición
              schema:
                type: string
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"
  /expenses:
    get:
      summary: Listar gastos
      operationId: listExpenses
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        "200":
          description: Lista de gastos
          headers:
            correlation-id:
              description: ID de correlación propagado en toda la petición
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      summary: Crear gasto
      operationId: createExpense
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: Clave idempotente única para reintentos seguros.
          schema:
            type: string
            minLength: 8
            maxLength: 255
        - name: correlation-id
          in: header
          required: false
          description: ID de correlación propagado end-to-end.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseRequest"
      responses:
        "201":
          description: Gasto creado
          headers:
            correlation-id:
              description: ID de correlación propagado en toda la petición
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-MCP-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterUserRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@planifi.app"
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
        fullName:
          type: string
          maxLength: 255
          example: "María López"
    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
    AuthResponse:
      type: object
      additionalProperties: false
      required:
        - userId
        - email
        - token
        - tokenType
        - expiresAt
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        token:
          type: string
        tokenType:
          type: string
          example: "Bearer"
        expiresAt:
          type: string
          format: date-time
    CreateExpenseRequest:
      type: object
      additionalProperties: false
      required:
        - amount
        - occurredOn
        - description
      properties:
        amount:
          type: number
          format: decimal
          example: 24.99
        occurredOn:
          type: string
          format: date
          example: "2025-01-15"
        description:
          type: string
          minLength: 1
          maxLength: 255
          example: "Taxi al aeropuerto"
    Expense:
      type: object
      additionalProperties: false
      required:
        - id
        - amount
        - occurredOn
        - description
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: number
          format: decimal
        occurredOn:
          type: string
          format: date
        description:
          type: string
        createdAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - errorCode
        - message
        - traceId
      properties:
        errorCode:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "amount must be greater than 0"
        traceId:
          type: string
          example: "9f2c3e3a3cdb4a2e"
    CreateApiKeyRequest:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "MCP principal"
    ApiKeySecret:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - apiKey
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        apiKey:
          type: string
          description: Secreto completo, visible solo en creación/rotación.
        createdAt:
          type: string
          format: date-time
  responses:
    BadRequestError:
      description: Solicitud inválida
      headers:
        correlation-id:
          description: ID de correlación propagado en toda la petición
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnauthorizedError:
      description: Credenciales inválidas o ausentes
      headers:
        correlation-id:
          description: ID de correlación propagado en toda la petición
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ConflictError:
      description: Conflicto por reutilización de Idempotency-Key con payload distinto
      headers:
        correlation-id:
          description: ID de correlación propagado en toda la petición
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Error interno
      headers:
        correlation-id:
          description: ID de correlación propagado en toda la petición
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFoundError:
      description: Recurso no encontrado
      headers:
        correlation-id:
          description: ID de correlación propagado en toda la petición
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"